# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Letters to God — Private Letters Vault API
  version: 0.1.0
  description: |
    REST API for the Private Letters Vault MVP.

    **Security:** Private letter title and body must never be sent in plaintext.
    All letter fields use structured encrypted payloads (see ADR-001 / ADR-003).

    **Auth:** Most routes require a NextAuth session cookie (sign in via the web app first).
    Use browser Swagger "Try it out" while logged in, or import the spec into Postman/Insomnia
    with your session cookie.

    **Spec file:** `docs/openapi.yaml` — served as JSON at `GET /api/openapi`.
servers:
  - url: http://localhost:3001
    description: Local development (default port)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        NextAuth JWT session cookie. Name may be `__Secure-next-auth.session-token` in production HTTPS.
  schemas:
...
    TwoFactorVerifyRequest:
      type: object
      properties:
        code:
          type: string
          pattern: '^\\d{6}$'
        backupCode:
          type: string
          minLength: 8
          maxLength: 32
paths:
...
  /api/auth/login/start:
    post:
      tags: [Auth]
      summary: Start credentials login (password verification; optional 2FA challenge)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Login token or 2FA challenge
        '401':
          description: Invalid credentials
        '429':
          description: Rate limited
  /api/auth/login/verify-2fa:
    post:
      tags: [Auth]
      summary: Complete credentials login after TOTP or backup code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeToken]
              allOf:
                - $ref: '#/components/schemas/TwoFactorVerifyRequest'
                - type: object
                  properties:
                    challengeToken:
                      type: string
                      minLength: 16
      responses:
        '200':
          description: One-time login token for NextAuth
        '401':
          description: Invalid challenge or code
        '429':
          description: Rate limited
  /api/auth/login/verify-2fa-oauth:
    post:
      tags: [Auth]
      summary: Complete OAuth login when account 2FA is enabled
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorVerifyRequest'
      responses:
        '200':
          description: Session upgrade token
        '401':
          description: Invalid code
        '429':
          description: Rate limited
...